feat: Refactor config to use accessors (#2)
* feat: Refactor config to use accessors

Because config might be accessed from a thread in mutable form, the pending update and consumption of an update is guarded with a mutex.
We should access the accessor with const modifier to get the version that does not modify to prevent locking.
Locking happens at fetch and destruction of accessor class

Make sure config does not segfault if we request a non-existent config
we simply return a pointer to static object (that is not part of config and will not synchronize)

Add config tests and make sure they pass